fix(dashboard): SparkLine container dimensions + Brand Voice markdown rendering
Two cosmetic fixes from sentinel's bug audit:

1. SparkLine: change wrapper from <span> to <div> and add minWidth/minHeight
   constraints to prevent Recharts ResponsiveContainer from computing negative
   container dimensions.

2. Brand Voice: replace raw <pre> with the existing renderMarkdown() renderer
   so markdown syntax (bold, lists, headings, etc.) displays as formatted HTML
   instead of raw text. Extract renderMarkdown to shared lib/render-markdown.tsx
   for reuse across kb-view and organization-tab.
